Minor fixes. Geant4Interface example should work now
Also, all build should be passing.

Do note that cmake gives /fpp CMAKE_Fortran_FLAGS on ifort. This breaks build with error:
```
#error: comment doesn't end to the end of file.
```
Removing that option in cmake cache file or specifying that in compile options is required to build.

This may correspond to the `-std=legacy` option in original cmake file, but not supported on ifort.